search: remove dead structural search frontend code
